refactor: code split tabbar and searchbar
