+++ src/index.js
@@ -1,0 +1,2 @@
+export { default as SearchBar } from './SearchBar';
+export { default as TabBar } from './TabBar';
